requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.0

# Supabase
supabase==2.7.0

# Redis
redis[hiredis]==5.0.7

# Azure Storage
azure-storage-blob==12.19.0
azure-identity==1.16.0

# Google Cloud
google-cloud-speech==2.26.0
google-auth==2.29.0

# LangChain & Gemini
langchain==0.2.0
langchain-google-genai==1.0.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# RevenueCat / HTTP
requests==2.31.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
httpx==0.27.0

# Audio processing
pydub==0.25.1

# Observability
newrelic==10.4.0
